+++ src/ProjectTemplate/Domains/.template/feature/CreateSample.Endpoint.cs
@@ -8,7 +8,7 @@ public partial class CreateSample
         {
             public void MapEndpoint(IEndpointRouteBuilder builder)
             {
-//#if (isCommand)
+//#if (isPost)
                 builder.MapPost("/sample", async (CreateSampleRequest request, IMediator mediator, CancellationToken cancellationToken) =>
                 {
                     return (await mediator.CreateSample(request, cancellationToken)).ToMinimalApiResult();
@@ -19,6 +19,42 @@ public partial class CreateSample
                     .Produces<ICreateSampleResponse>(StatusCodes.Status201Created)
                     .ProducesValidationProblem(StatusCodes.Status400BadRequest)
                     .Accepts<ICreateSampleRequest>("application/json")
+//#elseif (isPut)
+                builder.MapPut("/sample/{id:int}", async (int id, CreateSampleRequest request, IMediator mediator, CancellationToken cancellationToken) =>
+                {
+                    request.Id = id;
+                    return (await mediator.CreateSample(request, cancellationToken)).ToMinimalApiResult();
+                })
+                    .WithName("UpdateSample")
+                    .WithSummary("Updating a Sample")
+                    .WithDescription("This is a description on how a sample gets updated")
+                    .Produces<ICreateSampleResponse>(StatusCodes.Status200OK)
+                    .ProducesValidationProblem(StatusCodes.Status400BadRequest)
+                    .Produces(StatusCodes.Status404NotFound)
+                    .Accepts<ICreateSampleRequest>("application/json")
+//#elseif (isPatch)
+                builder.MapPatch("/sample/{id:int}", async (int id, CreateSampleRequest request, IMediator mediator, CancellationToken cancellationToken) =>
+                {
+                    request.Id = id;
+                    return (await mediator.CreateSample(request, cancellationToken)).ToMinimalApiResult();
+                })
+                    .WithName("PatchSample")
+                    .WithSummary("Patching a Sample")
+                    .WithDescription("This is a description on how a sample gets partially updated")
+                    .Produces<ICreateSampleResponse>(StatusCodes.Status200OK)
+                    .ProducesValidationProblem(StatusCodes.Status400BadRequest)
+                    .Produces(StatusCodes.Status404NotFound)
+                    .Accepts<ICreateSampleRequest>("application/json")
+//#elseif (isDelete)
+                builder.MapDelete("/sample/{id:int}", async (int id, IMediator mediator, CancellationToken cancellationToken) =>
+                {
+                    return (await mediator.CreateSample(new CreateSampleRequest { Id = id }, cancellationToken)).ToMinimalApiResult();
+                })
+                    .WithName("DeleteSample")
+                    .WithSummary("Deleting a Sample")
+                    .WithDescription("This is a description on how a sample gets deleted")
+                    .Produces(StatusCodes.Status204NoContent)
+                    .Produces(StatusCodes.Status404NotFound)
 //#else
                 builder.MapGet("/sample/{id:int}", async (int id, IMediator mediator, CancellationToken cancellationToken) =>
                 {
@@ -37,8 +73,14 @@ public partial class CreateSample
                     .AllowAnonymous()
                     .HasApiVersion(1, 0)
                     .Stable() // .Experimental() // .Deprecated() // .Hidden()
-//#if (isCommand)
+//#if (isPost)
                     .WithTags("Sample", "Create");
+//#elseif (isPut)
+                    .WithTags("Sample", "Update");
+//#elseif (isPatch)
+                    .WithTags("Sample", "Patch");
+//#elseif (isDelete)
+                    .WithTags("Sample", "Delete");
 //#else
                     .WithTags("Sample", "Get");
 //#endif